Show channel nicks & nick info on top of new backend.
ServerNicks is used to stored Nick Info and ChannelNicks is used to store nicks in a channel.

These two collections are global (for all Waartaa users). Since no fields in UserChannels collection
are updated to update Nick info, it does away with the jerks in the chat page due to re rendering
of templates dependent on UserChannel instances.

In the future, we'll try to poll for WHODATA for a server channel globally rather than per user.
This will reduce the number of writes happening to the database.