Fix loading older logs in chat room.
When the data-* attributes in HTML nodes are updated due to reactivity,
JS still reads the stale (initially rendered values of data-*). So, when we
were trying to load older logs of the current chat room, it'd actually
try to load the older logs from the chatroom where we had first tried
to load older logs. As a result, no change was reflected in the logs
of the currently selected chatroom. Reading the chatroom, server ids
from using .attr('data-foo') is able to read the up to date value
of the data attribute in HTML. Thus, it resolves this issue.